Se crea estructura de carpetas para el front con bloque de Ceiba

+++ nomina-back/src/main/java/com/ceiba/nomina/aplicacion/comando/manejador/ManejadorConsultarSalario.java
@@ -13,7 +13,7 @@ public class ManejadorConsultarSalario {
         this.servicioConsultarSalario = servicioConsultarSalario;
     }
 
-    public SalarioDto ejecutar(int idEmpleado){
+    public SalarioDto ejecutar(Long idEmpleado){
         return servicioConsultarSalario.ejecutar(idEmpleado);
     }
 }
