+++ nomina-back/src/main/java/com/ceiba/nomina/dominio/puerto/repositorio/RepositorioSalario.java
@@ -1,8 +1,9 @@
 package com.ceiba.nomina.dominio.src.main.java.com.ceiba.puerto.repositorio;
 
 import com.ceiba.nomina.dominio.src.main.java.com.ceiba.modelo.dto.SalarioDto;
+import java.lang.*;
 
 public interface RepositorioSalario {
 
-    SalarioDto consultar(int idEmpleado);
+    SalarioDto consultar(Long idEmpleado);
 }
